UndeadLink: fixes #3791 DirectHTTP: fixes #3785 GeneralSettings: updated
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@15034 ebf7c1c2-ba36-0410-9fe8-c592906822b4

+++ src/org/jdownloader/settings/GeneralSettings.java
@@ -216,9 +216,9 @@ public interface GeneralSettings extends ConfigInterface {
     @AboutConfig
     @Description("Penaltytime before a retry if JDownloader lost connection")
     @DefaultIntValue(5 * 60 * 1000)
-    long getWaittimeOnConnectionLoss();
+    int getWaittimeOnConnectionLoss();
 
-    void setWaittimeOnConnectionLoss(long milliseconds);
+    void setWaittimeOnConnectionLoss(int milliseconds);
 
     public static final KeyHandler DOWNLOAD_SPEED_LIMIT_ENABLED = CFG.getStorageHandler().getKeyHandler("DownloadSpeedLimitEnabled");
